fixup! Implement coinbase payout softfork
This fixes `getblocktemplate` RPC call which also needs
to pass consensus rules. Before, the consensus check
would fail because the softfork enforces a share of 99%
of the block reward to go to the whitelisted address
while a dummy script was used instead.

Now, miners can successfully use the RPC interface for mining.